# openapi.yaml
openapi: 3.1.0
info:
  title: Linewarmer Interview API
  version: 1.0.0
  summary: The backend contract expected by the Linewarmer frontend.
  description: |
    This document specifies the backend that `frontend/src/services/api.ts` expects.

    ## How this was derived

    The frontend talks to exactly one boundary: the `InterviewApi` interface in
    `frontend/src/services/api.ts`. Today that interface is satisfied by an in-memory
    mock (`frontend/src/services/mock/mockApi.ts`) wired up in
    `frontend/src/services/index.ts`; there is no HTTP client in the repository yet.

    Every operation below maps 1:1 to a method on `InterviewApi`. The `x-client-method`
    extension on each operation names the method it backs. Request and response schemas
    are the TypeScript types in `frontend/src/services/types.ts`, and the documented
    behaviour (filtering, sort order, state transitions, error text) is the observable
    behaviour of the mock, which is the de facto specification the UI was written against.

    ## Assumptions, flagged

    Because the frontend currently calls an in-process mock, a few things are not
    observable from the client code and are specified here as deliberate choices. A
    backend implementer should treat these as decisions to confirm, not as facts read
    off the frontend:

    1. **Transport of credentials.** The mock keeps `currentUserId` in a field and
       identifies guests by an opaque `participantId` held in `sessionStorage`
...

...

...

...

...

...
servers:
...
security:
  - interviewerSession: []
paths:
...
  /auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      x-client-method: InterviewApi.getCurrentUser
      summary: Get the signed-in interviewer, or null.
      description: |
        Returns the user identified by the request credentials, or the JSON literal
        `null` when there is no valid session. **Must not 401** — the client signature is
        `Promise<User | null>` and the app uses a null body as "signed out", so an error
        status would surface as a thrown exception instead.

        Guest participants are not users; this returns `null` for them.
      security:
        - interviewerSession: []
        - {}
      responses:
        '200':
          description: The current user, or null if unauthenticated.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/User'
                  - type: 'null'
              examples:
                signedIn:
                  value:
                    id: user_owner
                    email: jordan@linewarmer.io
                    displayName: Jordan Reyes
                    createdAt: '2026-09-09T10:00:00.000Z'
                signedOut:
                  value: null
  /auth/sign-in:
    post:
      tags: [auth]
      operationId: signIn
      x-client-method: InterviewApi.signIn
      summary: Sign in by email, creating the user on first sight.
      description: |
        Upsert semantics, mirroring the mock: if no user exists with this email, create
        one with `displayName` defaulting to the local part of the address
        (`email.split("@")[0]`), then sign that user in. The frontend never sends a
        password and has no sign-up screen, so a magic-link or OTP flow would need an
        extra round trip the client does not currently make.

        Establishes the `interviewerSession` credential on success.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  examples: ['jordan@linewarmer.io']
      responses:
        '200':
          description: Signed in. Sets the session cookie.
          headers:
            Set-Cookie:
              description: 'Session cookie, e.g. `lw_session=…; HttpOnly; Secure; SameSite=Lax; Path=/`'
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /sessions/{sessionId}/guest-links:
    parameters:
      - $ref: '#/components/parameters/SessionId'

    get:
      tags: [guest-links]
      operationId: listGuestLinks
      x-client-method: InterviewApi.listGuestLinks
      summary: List active invite links.
      description: |
        Owner only. Returns **only non-revoked** links (`revokedAt === null`). The client
        still re-checks `!l.revokedAt` when picking a link to copy, but it relies on this
        filter to keep dead tokens out of the share sheet.
      responses:
        '200':
          description: Active links for this session.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/GuestLink' }
        '403': { $ref: '#/components/responses/OwnerOnly' }
        '404': { $ref: '#/components/responses/SessionNotFound' }

    post:
      tags: [guest-links]
      operationId: createGuestLink
      x-client-method: InterviewApi.createGuestLink
      summary: Mint an invite link, rotating any existing one for that role.
      description: |
        Owner only. **Rotation, not accumulation:** creating a link revokes every other
        active link on this session with the same `roleGranted`, then issues a new one.
        There is at most one live link per role at a time. The room UI presents this as
        "rotate link", and any previously shared URL must stop working immediately.

        Defaults applied by the server: `roleGranted` defaults to `candidate` when the
        body is omitted (the client's parameter is optional), `maxUses: 10`,
        `expiresAt: null`, `uses: 0`.

        `token` must be a cryptographically random value of at least 128 bits. The
        frontend embeds it in a URL as `/join/{token}`.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  $ref: '#/components/schemas/GuestRole'
                  default: candidate
      responses:
        '201':
          description: The new link. Previous links for this role are now revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuestLink' }
        '403': { $ref: '#/components/responses/OwnerOnly' }
        '404': { $ref: '#/components/responses/SessionNotFound' }
  /sessions/{sessionId}/guest-links/{linkId}:
    parameters:
      - $ref: '#/components/parameters/SessionId'
      - name: linkId
        in: path
        required: true
        schema: { type: string, examples: ['lnk_m1a2b3c4d5'] }

    delete:
      tags: [guest-links]
      operationId: revokeGuestLink
      x-client-method: InterviewApi.revokeGuestLink
      summary: Revoke an invite link.
      description: |
        Owner only. Soft revoke — stamps `revokedAt` rather than deleting the row, so
        already-joined participants keep their access while the token stops working for
        new joins. The link must belong to `{sessionId}`; a mismatched pair is a 404.

        The room's "revoke all" control loops over active links calling this once each,
        so it must tolerate being called in quick succession.
      responses:
        '204':
          description: Revoked.
        '403': { $ref: '#/components/responses/OwnerOnly' }
        '404':
          description: No such link on this session.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { code: not_found, message: Link not found }

  # ---------------------------------------------------------- participants --
  /sessions/{sessionId}/participants:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      tags: [participants]
      operationId: listParticipants
      x-client-method: InterviewApi.listParticipants
      summary: List active participants.
      description: |
        Readable by any active participant of the session, including guests. Returns
        **only participants who have not left** (`leftAt === null`).

        The room route calls this on mount to decide whether the visitor is already a
        known participant before falling back to `joinAsOwner`.
      security:
        - interviewerSession: []
        - participantToken: []
      responses:
        '200':
          description: Active participants.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Participant' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/SessionNotFound' }
  /sessions/{sessionId}/participants/me:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    post:
      tags: [participants]
      operationId: joinAsOwner
      x-client-method: InterviewApi.joinAsOwner
      summary: Join a room as a signed-in interviewer.
      description: |
        The authenticated counterpart to `POST /join`. Requires an interviewer session;
        no guest token is involved.

        **Idempotent** — if the caller already has an active participant row on this
        session, return it unchanged rather than creating a second one. Re-entering a
        room must not duplicate the user in the roster.

        Role assignment: `owner` if `session.ownerUserId` is the caller, otherwise
        `interviewer`. `displayName` is taken from the user record. `color` is assigned
        round-robin from the participant palette.
      responses:
        '200':
          description: The session and the caller's participant record.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JoinResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/SessionNotFound' }
  /sessions/{sessionId}/participants/{participantId}:
    parameters:
      - $ref: '#/components/parameters/SessionId'
      - name: participantId
        in: path
        required: true
        schema: { type: string, examples: ['pt_m1a2b3c4d5'] }
    delete:
      tags: [participants]
      operationId: removeParticipant
      x-client-method: InterviewApi.removeParticipant
      summary: Remove someone from the room.
      description: |
        Owner only. Soft removal — stamps `leftAt`, drops the participant from presence,
        and revokes their access: their token must stop working for room reads and the
        WebSocket immediately.

        **Side effect:** broadcasts `presence_update` with the shortened roster.
      responses:
        '204':
          description: Removed.
        '403': { $ref: '#/components/responses/OwnerOnly' }
        '404':
          description: No such participant.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { code: not_found, message: Participant not found }
  /join:
    post:
      tags: [participants]
      operationId: joinWithToken
      x-client-method: InterviewApi.joinWithToken
      summary: Redeem a guest link and join as candidate or observer.
      description: |
        **Unauthenticated** — this is how a candidate with a shared URL enters. The token
        is the only credential, and it is spent here: on success the response establishes
        the `participantToken` credential used for every subsequent room call.

        Validation, in the order the mock applies it (order matters, because the first
        failure's message is what the candidate sees):

        | Check | Message |
        |---|---|
        | Token unknown | `This link is not valid.` |
        | `revokedAt` set | `This link has been revoked by the interviewer.` |
        | `expiresAt` in the past | `This link has expired.` |
        | Session archived | `This interview is no longer available.` |
        | Session ended | `This interview has ended.` |
        | Active participants >= `maxUses` | `This interview is full.` |
        | `displayName` under 2 chars after trimming | `Please enter your name.` |

        On success: increment the link's `uses`, create a participant with
        `role = link.roleGranted` and `userId: null`, `displayName` trimmed and truncated
        to 40 characters, and a palette colour.

        Note the capacity check counts *currently active participants*, not cumulative
        `uses` — someone who leaves frees a seat.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, displayName]
              properties:
                token:
                  type: string
                  description: The opaque token from the `/join/{token}` URL.
                displayName:
                  type: string
                  minLength: 2
                  maxLength: 40
                  description: Trimmed, then truncated to 40 characters.
                  examples: ['Sam Okafor']
      responses:
        '200':
          description: Joined. Issues the participant credential.
          headers:
            Set-Cookie:
              description: Optional participant cookie, if not returning a bearer token.
              schema: { type: string }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/JoinResult'
                  - type: object
                    properties:
                      participantToken:
                        type: string
                        description: |
                          Bearer credential scoped to this participant and session.
                          Not consumed by the frontend today — see assumption 1.
        '400':
          description: Name missing or too short.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { code: invalid_display_name, message: Please enter your name. }
        '403':
          description: Link revoked, expired, or unusable.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                revoked:
                  value: { code: link_revoked, message: This link has been revoked by the interviewer. }
                expired:
                  value: { code: link_expired, message: This link has expired. }
                ended:
                  value: { code: session_ended, message: This interview has ended. }
                archived:
                  value: { code: session_archived, message: This interview is no longer available. }
        '404':
          description: Unknown token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { code: link_invalid, message: This link is not valid. }
        '409':
          description: Session at capacity.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { code: session_full, message: This interview is full. }

  # ---------------------------------------------------------------- canvas --
  /sessions/{sessionId}/canvas:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      tags: [canvas]
      operationId: getCanvas
      x-client-method: InterviewApi.getCanvas
      summary: Fetch the current canvas snapshot.
      description: |
        Readable by any active participant. Returns the materialised document — every
        live element plus the monotonic `cursor` marking the last applied operation.

        A client that has a snapshot at cursor N and then receives `document_update`
        envelopes with cursors N+1, N+2 … is fully converged; this is the anchor the
        realtime channel builds on.
      security:
        - interviewerSession: []
        - participantToken: []
      responses:
        '200':
          description: The snapshot.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CanvasSnapshot' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/SessionNotFound' }
  /sessions/{sessionId}/canvas/clear:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    post:
      tags: [canvas]
      operationId: clearCanvas
      x-client-method: InterviewApi.clearCanvas
      summary: Clear the whole canvas.
      description: |
        Owner only, even though it takes an `actorId` — the mock calls `requireOwner`
        and then attributes the operation to the passed participant. Modelled as POST
        rather than DELETE because it carries a body.

        This is **not** a separate mutation path: it commits a normal `{ type: "clear" }`
        operation through the same log as realtime edits, advancing the cursor and
        fanning out a `document_update` to everyone in the room. Clients converge on it
        exactly like any other edit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actorId]
              properties:
                actorId:
                  type: string
                  description: |
                    Participant id to attribute the clear to — the caller's own
                    participant id, not their user id.
                  examples: ['pt_m1a2b3c4d5']
      responses:
        '204':
          description: Cleared. A `document_update` has been broadcast.
        '403': { $ref: '#/components/responses/OwnerOnly' }
        '404': { $ref: '#/components/responses/SessionNotFound' }

  # -------------------------------------------------------------- realtime --
...
components:
  securitySchemes:
    interviewerSession:
      type: apiKey
      in: cookie
      name: lw_session
      description: |
        Interviewer identity, established by `POST /auth/sign-in` and cleared by
        `POST /auth/sign-out`. Identifies a `User`, which is what owner checks resolve
        against. See assumption 1 — the mock has no wire credential, so the cookie is
        specified here rather than observed.
    participantToken:
      type: http
      scheme: bearer
      description: |
        Guest credential issued by `POST /join`, scoped to one participant in one
        session. Carries no user identity. Must stop working once the participant's
        `leftAt` is set. See assumption 1: the frontend currently persists only the
        `participantId` in `sessionStorage` and would need a small change to send this.
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      description: Session identifier, prefixed `ses_`.
      schema: { type: string, examples: ['ses_ratelimiter'] }
  responses:
    BadRequest:
      description: Malformed or invalid request body.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: No valid interviewer session.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: unauthenticated, message: Not authenticated }
    Forbidden:
      description: Authenticated, but not a member of this session.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: forbidden, message: 'Forbidden' }
    OwnerOnly:
      description: |
        The caller is not the session owner. Guarded by `requireOwner` in the mock, which
        compares `session.ownerUserId` against the current user.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: forbidden, message: 'Forbidden: owner only' }
    SessionNotFound:
      description: No session with that id.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: not_found, message: Session not found }
  schemas:
    Error:
      type: object
      required: [code, message]
      description: |
        Uniform error body. `message` is **user-facing copy** — the UI prints it
        verbatim (see `frontend/src/routes/join.$token.tsx`). `code` is for programmatic
        handling and is never displayed.
      properties:
        code:
          type: string
          description: Stable machine-readable identifier.
          examples: ['forbidden']
        message:
          type: string
          description: Human-readable sentence, shown directly to the end user.
          examples: ['This link has expired.']
    Role:
      type: string
      enum: [owner, interviewer, candidate, observer]
      description: |
        `owner` created the session. `interviewer` is a signed-in collaborator.
        `candidate` may edit only while `candidateEditingEnabled`. `observer` is
        always read-only.
    GuestRole:
      type: string
      enum: [candidate, observer]
      description: The subset of roles a guest link can grant.
    SessionState:
      type: string
      enum: [draft, live, ended, archived]
    User:
      type: object
      required: [id, email, displayName, createdAt]
      properties:
        id: { type: string, examples: ['user_owner'] }
        email: { type: string, format: email }
        displayName:
          type: string
          description: Defaults to the local part of the email at first sign-in.
        createdAt: { type: string, format: date-time }
    InterviewSession:
      type: object
      required:
        [id, ownerUserId, title, prompt, state, candidateEditingEnabled,
         cursorsVisible, scheduledAt, startedAt, endedAt, createdAt, updatedAt]
      properties:
        id: { type: string, examples: ['ses_ratelimiter'] }
        ownerUserId: { type: string }
        title: { type: string, examples: ['Design global rate limiter'] }
        prompt:
          type: string
          description: The question shown to the candidate. May be empty.
        state: { $ref: '#/components/schemas/SessionState' }
        candidateEditingEnabled:
          type: boolean
          description: |
            Whether participants with role `candidate` may write. Forced to `false` when
            the session ends. Enforced server-side on every incoming op — the client's
            own check is a UI affordance only.
        cursorsVisible:
          type: boolean
          description: Client-side rendering toggle for remote cursors.
        scheduledAt: { type: [string, 'null'], format: date-time }
        startedAt:
          type: [string, 'null']
          format: date-time
          description: Set once, on first start; not reset by a later start.
        endedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt:
          type: string
          format: date-time
          description: Refreshed on every mutation. Sort key for `GET /sessions`.
...
    GuestLink:
      type: object
      required:
        [id, sessionId, token, roleGranted, expiresAt, maxUses, uses, revokedAt, createdAt]
      properties:
        id: { type: string, examples: ['lnk_m1a2b3c4d5'] }
        sessionId: { type: string }
        token:
          type: string
          description: |
            Opaque, at least 128 bits of entropy (the mock emits 16 random bytes as 32
            hex characters). Embedded in the shared URL as `/join/{token}`.
          examples: ['9f2c1ab47e5d40318c6b0a7d2e5f1b93']
        roleGranted: { $ref: '#/components/schemas/GuestRole' }
        expiresAt:
          type: [string, 'null']
          format: date-time
          description: Null means no expiry. The mock never sets one.
        maxUses:
          type: [integer, 'null']
          description: |
            Capacity cap, defaulting to 10. Compared against the count of *currently
            active* participants, so it behaves as a concurrent-seat limit rather than a
            lifetime redemption count. Null means unlimited.
        uses:
          type: integer
          description: Cumulative successful redemptions.
        revokedAt:
          type: [string, 'null']
          format: date-time
          description: Soft-revocation stamp. Revoked links stay queryable but unusable.
        createdAt: { type: string, format: date-time }
    Participant:
      type: object
      required: [id, sessionId, userId, displayName, role, color, joinedAt, leftAt]
      properties:
        id: { type: string, examples: ['pt_m1a2b3c4d5'] }
        sessionId: { type: string }
        userId:
          type: [string, 'null']
          description: The signed-in user, or null for guests who joined by token.
        displayName: { type: string, maxLength: 40 }
        role: { $ref: '#/components/schemas/Role' }
        color:
          type: string
          enum: [amberdeep, warm, remote, live]
          description: |
            Palette key assigned round-robin on join, used to tint the participant's
            cursor and selection outline.

            **Not a free-form colour.** The value is a CSS custom property name minus its
            `--` prefix: the client interpolates it directly as `var(--${p.color})`
            (`frontend/src/components/interview/CanvasStage.tsx`). Only the four keys in
            this enum are defined in `frontend/src/styles.css`; any other value silently
            resolves to an invalid `var()` and renders uncoloured. Assign round-robin in
            the order listed, matching `PARTICIPANT_COLORS` in
            `frontend/src/services/mock/store.ts`.

            Note `amberdeep` is one token (a darker amber, distinct from `amber`), not
            two. `amber` itself is deliberately absent from this palette — it is reserved
            for highlighter strokes.
          examples: ['warm']
        joinedAt: { type: string, format: date-time }
        leftAt:
          type: [string, 'null']
          format: date-time
          description: |
            Soft-leave stamp. Non-null participants are excluded from
            `GET /sessions/{sessionId}/participants` and lose room access.
    JoinResult:
      type: object
      required: [session, participant]
      description: What both join paths return — the room to enter and who you are in it.
      properties:
        session: { $ref: '#/components/schemas/InterviewSession' }
        participant: { $ref: '#/components/schemas/Participant' }

    # ---- canvas document ----
...
    NodeElement:
      allOf:
        - $ref: '#/components/schemas/BaseElement'
        - type: object
          required: [kind, componentType, label, description, x, y, width, height]
          properties:
            kind: { const: node }
            componentType:
              type: string
              description: Key into the client's component library (e.g. cache, queue, db).
            label: { type: string }
            description: { type: string }
            x: { type: number }
            y: { type: number }
            width: { type: number }
            height: { type: number }
    StickyElement:
      allOf:
        - $ref: '#/components/schemas/BaseElement'
        - type: object
          required: [kind, text, x, y, width, height]
          properties:
            kind: { const: sticky }
            text: { type: string }
            x: { type: number }
            y: { type: number }
            width: { type: number }
            height: { type: number }
    TextElement:
      allOf:
        - $ref: '#/components/schemas/BaseElement'
        - type: object
          required: [kind, text, x, y]
          properties:
            kind: { const: text }
            text: { type: string }
            x: { type: number }
            y: { type: number }
    StrokeElement:
      allOf:
        - $ref: '#/components/schemas/BaseElement'
        - type: object
          required: [kind, points, color, width, highlighter]
          properties:
            kind: { const: stroke }
            points:
              type: array
              description: Flat interleaved coordinates — [x0, y0, x1, y1, …].
              items: { type: number }
            color:
              type: string
              enum: [ink, amber, remote, warm]
            width: { type: number }
            highlighter:
              type: boolean
              description: Renders semi-transparent and behind other elements.
...
    CanvasElement:
      oneOf:
        - $ref: '#/components/schemas/NodeElement'
        - $ref: '#/components/schemas/StickyElement'
        - $ref: '#/components/schemas/TextElement'
        - $ref: '#/components/schemas/StrokeElement'
        - $ref: '#/components/schemas/ConnectorElement'
      discriminator:
        propertyName: kind
        mapping:
          node: '#/components/schemas/NodeElement'
          sticky: '#/components/schemas/StickyElement'
          text: '#/components/schemas/TextElement'
          stroke: '#/components/schemas/StrokeElement'
          connector: '#/components/schemas/ConnectorElement'
...
    CanvasSnapshot:
      type: object
      required: [sessionId, cursor, elements, updatedAt]
      properties:
        sessionId: { type: string }
        cursor:
          type: integer
          description: Sequence number of the last operation folded into `elements`.
        elements:
          type: array
          items: { $ref: '#/components/schemas/CanvasElement' }
        updatedAt:
          type: string
          format: date-time
          description: ISO-8601 — unlike the numeric `updatedAt` on elements.

    # ---- realtime messages ----
...